# routes/admin/v1/admin.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FG Digital",
        "description": "ADMIN APIs - Project Management Portal",
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/admin/v1",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
...
    "paths": {
...
        "/plan/add": {
            "post": {
                "tags": ["3. Plan"],
                "summary": "Add New Plan",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["plan_name", "price", "days"],
                        "properties": {
                            "plan_name": {
                                "type": "string"
                            },
                            "price": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "days": {
                                "type": "integer",
                                "minimum": 1
                            }
                        },
                        "example": {
                            "plan_name": "Customize Plan X",
                            "price": "50000",
                            "days": "50"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/plan/update": {
            "post": {
                "tags": ["3. Plan"],
                "summary": "Update existing plan",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["plan_id"],
                        "properties": {
                            "plan_id": {
                                "type": "string"
                            },
                            "plan_name": {
                                "type": "string"
                            },
                            "price": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "days": {
                                "type": "integer",
                                "minimum": 1
                            }
                        },
                        "example": {
                            "plan_id": "-s65d51ve891s3",
                            "plan_name": "Customize Plan X",
                            "price": "50000",
                            "days": "50"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/plan/remove": {
            "get": {
                "tags": ["3. Plan"],
                "summary": "Remove Existing Plan",
                "parameters": [{
                    "in": "query",
                    "name": "plan_id",
                    "description": "Get Plan ID using /plan/get",
                    "required": true,
                    "items": {
                        "type": "string"
                    }
                }],
                "responses": {}
            }
        },
...
        "/client/create": {
            "post": {
                "tags": ["4. Client"],
                "summary": "Create New Client",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["email", "password", "name"],
                        "properties": {
                            "email": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "example": {
                            "email": "test@gmail.com",
                            "password": "my-secret-password",
                            "name": "Client XYZ"
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/client/update": {
            "post": {
                "tags": ["4. Client"],
                "summary": "Update existing client details",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["clientID"],
                        "properties": {
                            "clientID": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "example": {
                            "clientID": "-as65v16161as89ve",
                            "email": "test-new@gmail.com",
                            "name": "Client ABC"
                        }
                    }
                }],
                "responses": {}
            }
        },
...
        "/client/plan/add": {
            "post": {
                "tags": ["4. Client"],
                "summary": "Add new plan to client account",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "schema": {
                        "required": ["clientID", "plan", "start_date", "duration"],
                        "properties": {
                            "clientID": { "type": "string" },
                            "plan": { "type": "string", "description": "Get Plan Name using /plan/get" },
                            "start_date": { "type": "string", "format": "date" },
                            "duration": { "type": "integer" }
                        },
                        "example": {
                            "clientID": "-MCKMpR3RGE-EmVBF-JQ",
                            "plan": "Customized Plan",
                            "start_date": "2020/05/30",
                            "duration": 30
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/client/get": {
            "get": {
                "tags": ["4. Client"],
                "summary": "Get All Client",
                "parameters": [{
                    "in": "query",
                    "name": "client_id",
                    "description": "Client ID",
                    "required": false,
                    "items": { "type": "string" }
                }],
                "responses": {}
            }

        },
        "/client/change-password": {
            "post": {
                "tags": ["4. Client"],
                "summary": "Change password client",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "client_id and password",
                    "schema": {
                        "required": ["client_id"],
                        "properties": {
                            "client_id": { "type": "string" },
                            "password": { "type": "string" }
                        },
                        "example": {
                            "client_id": "-MDa5rMi4PeMnbevEOyV",
                            "password": "my-secret-password"
                        }
                    }
                }],
                "responses": {}
            }
        },
...
    }
}